Refactor into modular architecture with comprehensive docs
Split the monolithic lib.rs (640 lines) into focused Rust modules:
state, server, watchdog, setup/, platform/. Split App.tsx into
components (Mascot, StatusPill), hooks (useStatus, useDrag),
constants, types, and split CSS by concern.

Added architecture documentation in docs/ covering backend, frontend,
data flow, HTTP API, shell integration, setup flow, and storage plan.
