+++ src-tauri/src/helpers.rs
@@ -1,0 +1,19 @@
+use std::time::{SystemTime, UNIX_EPOCH};
+
+pub fn now_secs() -> u64 {
+    SystemTime::now()
+        .duration_since(UNIX_EPOCH)
+        .unwrap()
+        .as_secs()
+}
+
+pub fn get_query_param<'a>(url: &'a str, key: &str) -> Option<&'a str> {
+    let query = url.split('?').nth(1)?;
+    for pair in query.split('&') {
+        let mut kv = pair.splitn(2, '=');
+        if kv.next() == Some(key) {
+            return kv.next();
+        }
+    }
+    None
+}
