+++ src-tauri/src/platform/mod.rs
@@ -1,0 +1,1 @@
+pub mod macos;
