+++ src/styles/status-pill.css
@@ -1,56 +1,3 @@
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
-
-html,
-body {
-  height: 100%;
-  background: transparent;
-  overflow: hidden;
-}
-
-#root {
-  width: 100%;
-  height: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background: transparent;
-}
-
-.container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 4px;
-  cursor: grab;
-  user-select: none;
-  -webkit-user-select: none;
-}
-
-.container.dragging {
-  cursor: grabbing;
-}
-
-/* --- Sprite animation --- */
-.sprite {
-  image-rendering: pixelated;
-  background-repeat: no-repeat;
-  background-size: var(--sprite-width) 128px;
-  animation: sprite-play var(--sprite-duration) steps(var(--sprite-steps), end) infinite;
-}
-
-.sprite.frozen {
-  animation: none;
-}
-
-@keyframes sprite-play {
-  from { background-position: 0 0; }
-  to { background-position: calc(-1 * var(--sprite-width)) 0; }
-}
-
 /* --- Pill --- */
 .pill {
   display: inline-flex;
