+++ src/types/status.ts
@@ -1,0 +1,12 @@
+export type Status =
+  | "initializing"
+  | "searching"
+  | "idle"
+  | "busy"
+  | "service"
+  | "disconnected";
+
+export interface SpriteConfig {
+  file: string;
+  frames: number;
+}
